API_Clients/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy
pydantic>=1.10.7
pytest>=7.3.1
httpx>=0.24.0
python-dotenv
pyyaml
prometheus_client
python-jose[cryptography]
passlib[bcrypt]
prospector==1.3.1
redis>=4.5.4
psycopg2-binary
pydantic[email]
requests
pika